imagetoraster, imagetopdf: With print-scaling=none center image correctly
In the imageto... filters with print-scaling=none do the math for
centering the image on the page without the unprintable margins, so
that the image gets correctly centered relative to the page, not to
the printable area of the page.

As in imagetoraster we have to pass on a bitmap of the printable area
for CUPS Raster. we correct for the margins after calculating image
size and position without margins.

(manually backported from commit 0a1c76fc384615c8d22c171666a3157295bb789c)