pdftopdf: Fix N-up printing with long-edge-first
If the printer takes the paper long-edge-first and the number-up
option for printing multiple pages per sheet is used, the grid for the
shrinked pages onm the output sheet is set wrongly (number of lines
and number of columns swapped). This commit fixes the this bug.

(manually backported from commit 2477946645ee5511ce88c7e575ecc8334e2be1d3)